refactor(network): Prevent memory corruptions after sending a packet
When sending a packet we now copy the connection's scratch arena
locally (to a stack variable) instead of using `arena_save` and
`arena_restore`.

This led to memory corruption down the line, as the saved pointer is
set before to 0 (the beginning of the arena). When restoring to clean
the sending work, it would free ALL of the arena, counting stuff we
allocated to handle packets.

This caused a memory corruption after enabling compression and
acknowledging a player's authenticity.