feat(network,data): Start implementing dimension type registration
This is for RegistryData packets, because the vanilla client needs
them.

This consists of:
- Add codec functions for dimension types, notably from json and to
nbt. (The "to nbt" function is not finished yet)
- Define a struct for dimension types (DimensionType) and use it for
registry creation,
- Write a basic function to register dimension types (using codec
functions). This will later be generalized to other things.
- Add two registry functions: `registry_count` (returns the number of
things registered in a registry and `registry_foreach` (iterates over
all things registered inside a registry).
- Lay ground work to serialize dimension types into a packet.